pkg/component/extensions/operatingsystemconfig/original/components/varlibmount/component_test.go
// SPDX-FileCopyrightText: 2021 SAP SE or an SAP affiliate company and Gardener contributors
// SPDX-License-Identifier: Apache-2.0

package varlibmount_test

import (
	. "github.com/onsi/ginkgo/v2"
	. "github.com/onsi/gomega"
	"k8s.io/utils/pointer"

	extensionsv1alpha1 "github.com/gardener/gardener/pkg/apis/extensions/v1alpha1"
	"github.com/gardener/gardener/pkg/component/extensions/operatingsystemconfig/original/components"
	. "github.com/gardener/gardener/pkg/component/extensions/operatingsystemconfig/original/components/varlibmount"
)

var _ = Describe("Component", func() {
	Describe("#Config", func() {
		var component components.Component

		BeforeEach(func() {
			component = New()
		})

		It("should do nothing because kubelet data volume name is not set", func() {
			units, files, err := component.Config(components.Context{})

			Expect(err).NotTo(HaveOccurred())
			Expect(units).To(BeNil())
			Expect(files).To(BeNil())
		})

		It("should return the expected units and files", func() {
			units, files, err := component.Config(components.Context{KubeletDataVolumeName: pointer.String("foo")})

			Expect(err).NotTo(HaveOccurred())
			Expect(units).To(ConsistOf(
				extensionsv1alpha1.Unit{
					Name: "var-lib.mount",
					Content: pointer.String(`[Unit]
Description=mount /var/lib on kubelet data device
Before=kubelet.service
[Mount]
What=/dev/disk/by-label/kubeletdev
Where=/var/lib
Type=xfs
Options=defaults
[Install]
WantedBy=local-fs.target`),
				},
			))
			Expect(files).To(BeNil())
		})
	})
})
